lemrey fixes
