+++ samples/nrf9160/aws_fota/prj.conf
@@ -31,9 +31,6 @@ CONFIG_BSD_LIBRARY_SYS_INIT=n
 # Disable native network stack to save some memory
 CONFIG_NET_NATIVE=n
 
-CONFIG_NET_RX_STACK_SIZE=256
-CONFIG_NET_TX_STACK_SIZE=256
-
 # AT Host
 CONFIG_UART_INTERRUPT_DRIVEN=y
 CONFIG_AT_HOST_LIBRARY=y
