+++ subsys/dfu/include/dfu_target_modem.h
@@ -3,16 +3,20 @@
  *
  * SPDX-License-Identifier: LicenseRef-BSD-5-Clause-Nordic
  */
+
+/** @file dfu_target_modem.h
+ *
+ * @defgroup dfu_target_modem Modem DFU Target
+ * @{
+ * @brief DFU Target for upgrades performed by Modem
+ */
+
 #ifndef DFU_TARGET_MODEM_H__
 #define DFU_TARGET_MODEM_H__
-
-#include <zephyr/types.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-#include <dfu/dfu_target.h>
 
 /**
  * @brief See if data in buf indicates modem upgrade.
@@ -33,7 +37,7 @@ int dfu_target_modem_init(void);
  *
  * @return Offset of firmware
  */
-int dfu_target_modem_offset(void);
+int dfu_target_modem_offset(size_t *offset);
 
 /**
  * @brief Write firmware data.
@@ -51,17 +55,6 @@ int dfu_target_modem_write(const void *const buf, size_t len);
  * @return 0 on success, negative errno otherwise.
  */
 int dfu_target_modem_done(void);
-
-/** @brief Expose API compatible with dfu_target. This is used by
- *	   dfu_target.c.
- */
-static struct dfu_target dfu_target_modem = {
-	.identify = dfu_target_modem_identify,
-	.init = dfu_target_modem_init,
-	.offset = dfu_target_modem_offset,
-	.write = dfu_target_modem_write,
-	.done = dfu_target_modem_done,
-};
 
 #endif /* DFU_TARGET_MODEM_H__ */
